refactor dump management logic into own class
